Circular mean added
Since the angle could be close to 0 we need to do some math to get a meaningful average angle.

To do that we convert the angle to the x and y coordinates on the unit circle:
x=cos(angle)
y=sin(angle)

Where angle is in radians.  But I bet that the little esp does not implement Math.h  so this code will not work.