Update Main.ino
What a coincidence! A pointer is 4 bytes and a float is 4 bytes. But alas, a double is 8 bytes.  So when I switched stuff to use double instead of floats I found a bug.

double myArray[10];
myPointer pointyPointer = &myArray;
sizeof(myArray) = 80;
sizeof(pointyPointer) = 4;
So sizeof(pointyPointer)/sizeof(double) = 0 and averaging over 0 items is bad
But if myArray contains float then ( sizeof(pointyPointer)/sizeof(float) = 1
and averaging over one item actually is not fatal.